# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BOC Study Notebook API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notebooks/{id}/notes:
    get:
      operationId: listNotes
      tags: [notes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Note" }
    post:
      operationId: createNote
      tags: [notes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateNote
      tags: [notes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }
    delete:
      operationId: deleteNote
      tags: [notes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }

  # ---- Flashcards ----
...
  /quizzes:
    get:
      operationId: listQuizAttempts
      tags: [quizzes]
      summary: Recent practice quiz attempts
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, default: 20 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/QuizSummary" }
    post:
      operationId: startQuiz
      tags: [quizzes]
      summary: Start a practice quiz session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuizInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quiz" }
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quiz" }
        "404": { $ref: "#/components/responses/NotFound" }
  /quizzes/{id}/answer:
    post:
      operationId: answerQuizQuestion
      tags: [quizzes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuizAnswerInput" }
      responses:
        "200":
          description: OK with feedback
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizAnswerResult" }
  /quizzes/{id}/finish:
    post:
      operationId: finishQuiz
      tags: [quizzes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quiz" }

  # ---- Mock exam ----
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OpenaiConversation" }
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OpenaiConversationInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OpenaiConversation" }
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OpenaiConversationWithMessages" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OpenaiMessage" }
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send tutor message; SSE stream of response chunks
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OpenaiMessageInput" }
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream: {}

  # ---- Resources library ----
...
components:
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    ApiError:
      type: object
      properties: { error: { type: string } }
      required: [error]
...
    Note:
      type: object
      required: [id, notebookId, title, content, sourceKind, createdAt]
      properties:
        id: { type: integer }
        notebookId: { type: integer }
        title: { type: string }
        content: { type: string }
        sourceKind:
          type: string
          enum: [text, paste, url, pdf]
        sourceUrl: { type: ["string", "null"] }
        topicId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    NoteInput:
      type: object
      required: [title, content]
      properties:
        title: { type: string, minLength: 1 }
        content: { type: string }
        sourceKind:
          type: string
          enum: [text, paste, url, pdf]
        sourceUrl: { type: string }
        topicId: { type: integer }
    NoteUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        content: { type: string }
        topicId: { type: integer }
...
    QuizSummary:
      type: object
      required: [id, mode, totalQuestions, correctCount, finishedAt]
      properties:
        id: { type: integer }
        mode: { type: string, enum: [adaptive, topic, domain, weakness] }
        totalQuestions: { type: integer }
        correctCount: { type: integer }
        finishedAt: { type: ["string", "null"], format: date-time }
    Quiz:
      type: object
      required: [id, mode, questions, currentIndex, finished]
      properties:
        id: { type: integer }
        mode: { type: string, enum: [adaptive, topic, domain, weakness] }
        notebookId: { type: ["integer", "null"] }
        topicId: { type: ["integer", "null"] }
        domainId: { type: ["integer", "null"] }
        questions:
          type: array
          items: { $ref: "#/components/schemas/QuizQuestion" }
        currentIndex: { type: integer }
        finished: { type: boolean }
        score: { type: ["number", "null"] }
    QuizQuestion:
      type: object
      required: [id, questionId, stem, choices]
      properties:
        id: { type: integer }
        questionId: { type: integer }
        stem: { type: string }
        choices:
          type: array
          items: { type: string }
        topicId: { type: ["integer", "null"] }
        domainId: { type: ["integer", "null"] }
        selectedIndex: { type: ["integer", "null"] }
        correctIndex: { type: ["integer", "null"] }
        rationale: { type: ["string", "null"] }
        sourceUrl: { type: ["string", "null"] }
    QuizInput:
      type: object
      required: [mode, count]
      properties:
        mode: { type: string, enum: [adaptive, topic, domain, weakness] }
        count: { type: integer, minimum: 1, maximum: 50 }
        notebookId: { type: integer }
        topicId: { type: integer }
        domainId: { type: integer }
    QuizAnswerInput:
      type: object
      required: [questionId, selectedIndex]
      properties:
        questionId: { type: integer }
        selectedIndex: { type: integer, minimum: 0 }
    QuizAnswerResult:
      type: object
      required: [correct, correctIndex, rationale]
      properties:
        correct: { type: boolean }
        correctIndex: { type: integer }
        rationale: { type: string }
        sourceUrl: { type: ["string", "null"] }
...
    OpenaiConversation:
      type: object
      required: [id, title, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        notebookId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    OpenaiConversationInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        notebookId: { type: integer }
    OpenaiMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id: { type: integer }
        conversationId: { type: integer }
        role: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    OpenaiMessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
    OpenaiConversationWithMessages:
      type: object
      required: [id, title, createdAt, messages]
      properties:
        id: { type: integer }
        title: { type: string }
        notebookId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
        messages:
          type: array
          items: { $ref: "#/components/schemas/OpenaiMessage" }
...